Add error-mode handling in SQL execution
This is a bit of a hack, since we will swap out for a new SQL
execution in the near future.

We no longer wrap commands in transactions; that was getting in the
way of non-strict execution of batches of commands. Once there is an
error in a transaction, even one you would like to ignore, any further
commands automatically fail.

Basic tests added for "normal" mode.